restructured code to better separate workspacke and project implementation
